Refactored Code
So, I was hounded by a few people to put together a little Sudoku
program.  Dunno why, it's not like there aren't many out there for free.
It was rather fun, though.  This could do with a little future work to
ensure that on harder puzzles (17 revealed numbers, for example) there
is only ever one valid solution.  Still, quite fun to write.